docs: create story for NotificationList
