--- packages/main/src/webComponents/NotificationListItem/NotificationListItem.mdx
@@ -1,43 +1,0 @@
-import { ArgTypesWithNote, ControlsWithNote, DocsHeader, Footer } from '@sb/components';
-import SubcomponentsSection from '@sb/docs/SubcomponentsSection.md?raw';
-import { Canvas, Description, Markdown, Meta } from '@storybook/blocks';
-import { Menu } from '../Menu';
-import { MenuItem } from '../MenuItem';
-import * as ComponentStories from './NotificationListItem.stories';
-import { excludePropsForAbstract } from '@sb/utils';
-
-<Meta of={ComponentStories} />
-
-<DocsHeader since="0.10.0" />
-
-<br />
-
-## Example
-
-<Canvas of={ComponentStories.Default} />
-
-## Properties
-
-<ControlsWithNote of={ComponentStories.Default} />
-
-# More Examples
-
-<br />
-
-## List with NotificationListItems
-
-<Canvas of={ComponentStories.MultipleNotifications} />
-
-<Markdown>{SubcomponentsSection}</Markdown>
-
-## Menu
-
-<Description of={Menu} />
-<ArgTypesWithNote hideHTMLPropsNote exclude={excludePropsForAbstract} of={Menu} />
-
-## MenuItem
-
-<Description of={MenuItem} />
-<ArgTypesWithNote hideHTMLPropsNote exclude={excludePropsForAbstract} of={MenuItem} />
-
-<Footer />
